Crate task_struct

This crate contains the basic Task structure, which holds contextual execution states needed to support safe multithreading.

To create new Tasks, use the spawn crate. For more advanced task-related types, see the task crate.

Structs

PanicInfoOwned

Just like core::panic::PanicInfo, but with owned String types instead of &str references.

RestartInfo

A struct holding data items needed to restart a Task.

Task

A structure that contains contextual information for a thread of execution.

TaskInner

The parts of a Task that may be modified after its creation.

Enums

ExitValue

The two ways a Task can exit, including possible return values and conditions.

InheritedStates

The states used to initialize a new Task when creating it; see Task::new().

KillReason

The list of possible reasons that a given Task was killed prematurely.

RunState

The set of possible runstates that a Task can be in.

Type Definitions

KillHandler

The function signature of the callback that will be invoked when a Task panics or otherwise fails, e.g., a machine exception occurs.